Adds Devise and setups the basic Note model
Went through the basic Devise install steps to get the bare bone setup
to allow sign up and login for a user. I also created the simple Note
model that will be central for the note taking aspect of for this
applicaiton. Part of this work also included the basic setup for the
databases as well leveraging "dotenv-rails"